Cosmetic changes. Strategy Design pattern fix - moved from spring to pure strategy way. Local DB - Make DB folder if no exists - fix
Committer : Ronm
Reviewer :  daniels

+++ src/main/java/interpreter/MathExpression.java
@@ -1,5 +1,5 @@
 package interpreter;
 
 public interface MathExpression {
-	public int interpret();
+	public double interpret();
 }
